moving files around and removing unnecessary "organizations" files
