--- app/(host-dashboard)/host/my-organization/page.tsx
@@ -1,7 +1,0 @@
-export default function HostOrgPage() {
-  return (
-    <div className='flex flex-col gap-5'>
-      <div className='flex flex-row gap-5 '>A host can crud their organization</div>
-    </div>
-  )
-}
