--- app/(host-dashboard)/host/organizations/[orgSlug]/page.tsx
@@ -1,7 +1,0 @@
-export default function AdminOrganizationPage() {
-  return (
-    <div className='flex flex-col gap-5'>
-      <div className='flex flex-row gap-5 '>A admin can CRUD an organization</div>
-    </div>
-  )
-}
