--- app/(host-dashboard)/host/page.tsx
@@ -1,7 +1,0 @@
-export default function HostDashboardPage() {
-  return (
-    <div className='flex flex-col gap-5'>
-      <div className='flex flex-row gap-5 '>Host Dashboard</div>
-    </div>
-  )
-}
