--- app/(host-dashboard)/host/payments/page.tsx
@@ -1,7 +1,0 @@
-export default function HostPaymentPage() {
-  return (
-    <div className='flex flex-col gap-5'>
-      <div className='flex flex-row gap-5 '>A host can crud their payments</div>
-    </div>
-  )
-}
