--- app/(host-dashboard)/layout.tsx
@@ -1,345 +1,0 @@
-'use client'
-
-import { Fragment, useContext, useState } from 'react'
-import { GlobalContext } from '@/data/globalContext'
-import { Dialog, Menu, Transition } from '@headlessui/react'
-import {
-  Bars3Icon,
-  BellIcon,
-  BriefcaseIcon,
-  Cog6ToothIcon,
-  BuildingOfficeIcon,
-  XMarkIcon,
-  CurrencyDollarIcon,
-} from '@heroicons/react/24/outline'
-import { ChevronDownIcon, MagnifyingGlassIcon } from '@heroicons/react/20/solid'
-import { Switch } from '@/components/ui/switch'
-import { useRouter, usePathname } from 'next/navigation'
-import Link from 'next/link'
-import Logo from '@/components/shared/Logo'
-
-const navigation = [
-  { name: 'Properties', href: '/host/properties', icon: BriefcaseIcon },
-  {
-    name: 'My Organization',
-    href: '/host/my-organization',
-    icon: BuildingOfficeIcon,
-  },
-  { name: 'Payments', href: '/host/payments', icon: CurrencyDollarIcon },
-]
-
-const userNavigation = [
-  { name: 'Your profile', href: '#' },
-  { name: 'Sign out', href: '#' },
-]
-
-function classNames(...classes: any[]) {
-  return classes.filter(Boolean).join(' ')
-}
-
-const DashboardLayout = ({ children }: { children: React.ReactNode }) => {
-  const { updateState, darkMode } = useContext(GlobalContext)
-  const [sidebarOpen, setSidebarOpen] = useState(false)
-  const router = useRouter()
-  const currPath = usePathname()
-
-  return (
-    <div>
-      <div>
-        <Transition.Root show={sidebarOpen} as={Fragment}>
-          <Dialog
-            as='div'
-            className='relative z-50 lg:hidden'
-            onClose={setSidebarOpen}
-          >
-            <Transition.Child
-              as={Fragment}
-              enter='transition-opacity ease-linear duration-300'
-              enterFrom='opacity-0'
-              enterTo='opacity-100'
-              leave='transition-opacity ease-linear duration-300'
-              leaveFrom='opacity-100'
-              leaveTo='opacity-0'
-            >
-              <div className='fixed inset-0 bg-gray-900/80' />
-            </Transition.Child>
-
-            <div className='fixed inset-0 flex'>
-              <Transition.Child
-                as={Fragment}
-                enter='transition ease-in-out duration-300 transform'
-                enterFrom='-translate-x-full'
-                enterTo='translate-x-0'
-                leave='transition ease-in-out duration-300 transform'
-                leaveFrom='translate-x-0'
-                leaveTo='-translate-x-full'
-              >
-                <Dialog.Panel className='relative mr-16 flex w-full max-w-xs flex-1'>
-                  <Transition.Child
-                    as={Fragment}
-                    enter='ease-in-out duration-300'
-                    enterFrom='opacity-0'
-                    enterTo='opacity-100'
-                    leave='ease-in-out duration-300'
-                    leaveFrom='opacity-100'
-                    leaveTo='opacity-0'
-                  >
-                    <div className='absolute left-full top-0 flex w-16 justify-center pt-5'>
-                      <button
-                        type='button'
-                        className='-m-2.5 p-2.5'
-                        onClick={() => setSidebarOpen(false)}
-                      >
-                        <span className='sr-only'>Close sidebar</span>
-                        <XMarkIcon
-                          className='h-6 w-6 text-white'
-                          aria-hidden='true'
-                        />
-                      </button>
-                    </div>
-                  </Transition.Child>
-                  {/* Sidebar component, swap this element with another sidebar if you like */}
-                  <div className='flex grow flex-col gap-y-5 overflow-y-auto px-6 pb-4'>
-                    <div className='flex h-16 shrink-0 items-center'>
-                      <Logo src='' />
-                    </div>
-                    <nav className='flex flex-1 flex-col'>
-                      <ul role='list' className='flex flex-1 flex-col gap-y-7'>
-                        <li>
-                          <ul role='list' className='-mx-2 space-y-1'>
-                            {navigation.map((item) => (
-                              <li key={item.name}>
-                                <Link
-                                  href={item.href}
-                                  className={classNames(
-                                    currPath === item.href
-                                      ? 'bg-gray-50 text-primaryGreen dark:bg-gray-800'
-                                      : 'text-gray-700 hover:bg-gray-50 hover:text-primaryGreen hover:dark:bg-gray-800',
-                                    'group flex gap-x-3 rounded-md p-2 text-sm font-semibold leading-6'
-                                  )}
-                                >
-                                  <item.icon
-                                    className={classNames(
-                                      currPath === item.href
-                                        ? 'text-primaryGreen'
-                                        : 'text-gray-400 group-hover:text-primaryGreen',
-                                      'h-6 w-6 shrink-0'
-                                    )}
-                                    aria-hidden='true'
-                                  />
-                                  {item.name}
-                                </Link>
-                              </li>
-                            ))}
-                          </ul>
-                        </li>
-                        <li className='mt-auto'>
-                          <a
-                            href='#'
-                            className='group -mx-2 flex gap-x-3 rounded-md p-2 text-sm font-semibold leading-6 text-gray-700 hover:bg-gray-50 hover:text-primaryGreen'
-                          >
-                            <Cog6ToothIcon
-                              className='h-6 w-6 shrink-0 text-gray-400 group-hover:text-primaryGreen'
-                              aria-hidden='true'
-                            />
-                            Settings
-                          </a>
-                        </li>
-                      </ul>
-                    </nav>
-                  </div>
-                </Dialog.Panel>
-              </Transition.Child>
-            </div>
-          </Dialog>
-        </Transition.Root>
-
-        {/* Static sidebar for desktop */}
-        <div className='hidden lg:fixed lg:inset-y-0 lg:z-50 lg:flex lg:w-72 lg:flex-col'>
-          {/* Sidebar component, swap this element with another sidebar if you like */}
-          <div className='flex grow flex-col gap-y-5 overflow-y-auto border-r px-6 pb-4'>
-            <div className='flex h-16 shrink-0 items-center'>
-            <Logo src='' />
-            </div>
-            <nav className='flex flex-1 flex-col'>
-              <ul role='list' className='flex flex-1 flex-col gap-y-7'>
-                <li>
-                  <ul role='list' className='-mx-2 space-y-1'>
-                    {navigation.map((item) => (
-                      <li key={item.name}>
-                        <Link
-                          href={item.href}
-                          className={classNames(
-                            currPath === item.href
-                              ? 'bg-gray-50 text-primaryGreen dark:bg-gray-800'
-                              : 'text-gray-700 hover:bg-gray-50 hover:text-primaryGreen hover:dark:bg-gray-800',
-                            'group flex gap-x-3 rounded-md p-2 text-sm font-semibold leading-6'
-                          )}
-                        >
-                          <item.icon
-                            className={classNames(
-                              currPath === item.href
-                                ? 'text-primaryGreen'
-                                : 'text-gray-400 group-hover:text-primaryGreen',
-                              'h-6 w-6 shrink-0'
-                            )}
-                            aria-hidden='true'
-                          />
-                          {item.name}
-                        </Link>
-                      </li>
-                    ))}
-                  </ul>
-                </li>
-                <li className='mt-auto'>
-                  <Menu as='div' className='relative'>
-                    <Menu.Button className='-m-1.5 flex items-center p-1.5'>
-                      <a
-                        href='#'
-                        className='group -mx-2 flex gap-x-3 rounded-md p-2 pr-5 text-sm font-semibold leading-6 text-gray-700 hover:bg-gray-50 hover:text-primaryGreen hover:dark:bg-gray-800'
-                      >
-                        <Cog6ToothIcon
-                          className='h-6 w-6 shrink-0 text-gray-400 group-hover:text-primaryGreen'
-                          aria-hidden='true'
-                        />
-                        Settings
-                      </a>
-                    </Menu.Button>
-                    <Transition
-                      as={Fragment}
-                      enter='transition ease-out duration-100'
-                      enterFrom='transform opacity-0 scale-95'
-                      enterTo='transform opacity-100 scale-100'
-                      leave='transition ease-in duration-75'
-                      leaveFrom='transform opacity-100 scale-100'
-                      leaveTo='transform opacity-0 scale-95'
-                    >
-                      <Menu.Items className='absolute bottom-12 left-6 z-10 mt-2.5 origin-top-right rounded-md py-2 shadow-lg ring-1 ring-gray-900/5 focus:outline-none dark:bg-gray-800'>
-                        <Menu.Item key='darkModeSwitch'>
-                          <div className='flex flex-row gap-3 p-5'>
-                            <span>Dark Mode</span>
-                            <Switch
-                              checked={darkMode}
-                              onCheckedChange={(value) => {
-                                updateState({ darkMode: !!value })
-                              }}
-                            />
-                          </div>
-                        </Menu.Item>
-                      </Menu.Items>
-                    </Transition>
-                  </Menu>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-
-        <div className='lg:pl-72'>
-          <div className='sticky top-0 z-40 flex h-16 shrink-0 items-center gap-x-4 border-b px-4 shadow-sm sm:gap-x-6 sm:px-6 lg:px-8'>
-            <button
-              type='button'
-              className='-m-2.5 p-2.5 text-gray-700 lg:hidden'
-              onClick={() => setSidebarOpen(true)}
-            >
-              <span className='sr-only'>Open sidebar</span>
-              <Bars3Icon className='h-6 w-6' aria-hidden='true' />
-            </button>
-
-            {/* Separator */}
-            <div className='h-6 w-px lg:hidden' aria-hidden='true' />
-
-            <div className='flex flex-1 gap-x-4 self-stretch lg:gap-x-6'>
-              <form className='relative flex flex-1' action='#' method='GET'>
-                <label htmlFor='search-field' className='sr-only'>
-                  Search
-                </label>
-                <MagnifyingGlassIcon
-                  className='pointer-events-none absolute inset-y-0 left-0 h-full w-5 text-gray-400'
-                  aria-hidden='true'
-                />
-                <input
-                  id='search-field'
-                  className='block h-full w-full border-0 bg-background py-0 pl-8 pr-0 placeholder:text-gray-400 focus:ring-0 sm:text-sm'
-                  placeholder='Search...'
-                  type='search'
-                  name='search'
-                />
-              </form>
-              <div className='flex items-center gap-x-4 lg:gap-x-6'>
-                <button
-                  type='button'
-                  className='-m-2.5 p-2.5 text-gray-400 hover:text-gray-500'
-                >
-                  <span className='sr-only'>View notifications</span>
-                  <BellIcon className='h-6 w-6' aria-hidden='true' />
-                </button>
-
-                {/* Separator */}
-                <div
-                  className='hidden lg:block lg:h-6 lg:w-px lg:bg-gray-200'
-                  aria-hidden='true'
-                />
-
-                {/* Profile dropdown */}
-                <Menu as='div' className='relative'>
-                  <Menu.Button className='-m-1.5 flex items-center p-1.5'>
-                    <span className='sr-only'>Open user menu</span>
-                    <img
-                      className='h-8 w-8 rounded-full bg-gray-50'
-                      src='https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80'
-                      alt=''
-                    />
-                    <span className='hidden lg:flex lg:items-center'>
-                      <span
-                        className='ml-4 text-sm font-semibold leading-6'
-                        aria-hidden='true'
-                      >
-                        Tom Cook
-                      </span>
-                      <ChevronDownIcon
-                        className='ml-2 h-5 w-5 text-gray-400'
-                        aria-hidden='true'
-                      />
-                    </span>
-                  </Menu.Button>
-                  <Transition
-                    as={Fragment}
-                    enter='transition ease-out duration-100'
-                    enterFrom='transform opacity-0 scale-95'
-                    enterTo='transform opacity-100 scale-100'
-                    leave='transition ease-in duration-75'
-                    leaveFrom='transform opacity-100 scale-100'
-                    leaveTo='transform opacity-0 scale-95'
-                  >
-                    <Menu.Items className='absolute right-0 z-10 mt-2.5 w-32 origin-top-right rounded-md bg-white py-2 shadow-lg ring-1 ring-gray-900/5 focus:outline-none'>
-                      {userNavigation.map((item) => (
-                        <Menu.Item key={item.name}>
-                          {({ active }) => (
-                            <a
-                              href={item.href}
-                              className={classNames(
-                                active ? 'bg-gray-50' : '',
-                                'block px-3 py-1 text-sm leading-6 text-gray-900'
-                              )}
-                            >
-                              {item.name}
-                            </a>
-                          )}
-                        </Menu.Item>
-                      ))}
-                    </Menu.Items>
-                  </Transition>
-                </Menu>
-              </div>
-            </div>
-          </div>
-
-          <div className='container py-10'>{children}</div>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-export default DashboardLayout
