+++ components/ListPage.tsx
@@ -3,10 +3,10 @@
 import React, { Dispatch, SetStateAction, useContext, useEffect, useState } from 'react'
 import { useRouter } from 'next/navigation'
 import { Dialog, DialogContent } from "@/components/ui/dialog"
-import { DataTable, DataTableProps, FilterConfigType, TableColumns } from '../shared/table/DataTable'
-import { DataTablePageHeader, PageActionButtonType } from '../shared/table/DataTablePageHeader'
-import { DataTableDialogContent, DataTableDialogContentProps } from '../shared/table/DataTableDialogContent'
-import Breadcrumb from '../breadcrumb'
+import { DataTable, DataTableProps, FilterConfigType, TableColumns } from './table/DataTable'
+import { DataTablePageHeader, PageActionButtonType } from './table/DataTablePageHeader'
+import { DataTableDialogContent, DataTableDialogContentProps } from './table/DataTableDialogContent'
+import Breadcrumb from './breadcrumb'
 import { ChevronRight, Download } from 'lucide-react'
 import { ColumnFiltersState, PaginationState, SortingState } from '@tanstack/react-table'
 import { GlobalContext } from '@/data/globalContext'
