+++ data/testData/addresses.json
@@ -1,72 +1,1 @@
-[
-  {
-    "id": "BZE58966",
-    "street": "433 Greenholt Spurs",
-    "city": "Runteworth",
-    "state": "Connecticut",
-    "zipCode": "97732"
-  },
-  {
-    "id": "QAO87364",
-    "street": "542 Lake Avenue",
-    "city": "Port Hermanfort",
-    "state": "Kansas",
-    "zipCode": "24104-0748"
-  },
-  {
-    "id": "BZA57636",
-    "street": "760 7th Street",
-    "city": "Castle Rock",
-    "state": "Indiana",
-    "zipCode": "70954"
-  },
-  {
-    "id": "ZFX88165",
-    "street": "617 Main Avenue",
-    "city": "Fort Edgartown",
-    "state": "Vermont",
-    "zipCode": "14845"
-  },
-  {
-    "id": "VKB89750",
-    "street": "461 Alicia Loop",
-    "city": "East Jermeyborough",
-    "state": "Washington",
-    "zipCode": "85488-7067"
-  },
-  {
-    "id": "FRJ95095",
-    "street": "739 Kassulke Islands",
-    "city": "West Alifield",
-    "state": "Nebraska",
-    "zipCode": "61317-1759"
-  },
-  {
-    "id": "DEN46954",
-    "street": "2598 Sterling Forge",
-    "city": "Kearaland",
-    "state": "Hawaii",
-    "zipCode": "69856-1457"
-  },
-  {
-    "id": "SUK12348",
-    "street": "64968 Buckridge Harbors",
-    "city": "Port Abner",
-    "state": "West Virginia",
-    "zipCode": "42801"
-  },
-  {
-    "id": "BRA29753",
-    "street": "3720 Kozey Creek",
-    "city": "North Sedrick",
-    "state": "Hawaii",
-    "zipCode": "62328-1485"
-  },
-  {
-    "id": "TUZ29864",
-    "street": "8605 Pfeffer Spurs",
-    "city": "Lake Laisha",
-    "state": "Alaska",
-    "zipCode": "48308-3332"
-  }
-]
+[]
