+++ data/testData/contacts.json
@@ -1,62 +1,1 @@
-[
-  {
-    "id": "KFZ47645",
-    "name": "Kelvin Sipes",
-    "email": "Agustina.Lowe@hotmail.com",
-    "phone": "523.751.2604 x6499"
-  },
-  {
-    "id": "XEZ82429",
-    "name": "Esther Williamson II",
-    "email": "Murray.Kshlerin@hotmail.com",
-    "phone": "766-629-7221 x07598"
-  },
-  {
-    "id": "WVD55140",
-    "name": "Rolando Schaefer",
-    "email": "Verner24@yahoo.com",
-    "phone": "730.679.7251"
-  },
-  {
-    "id": "BRH76774",
-    "name": "Herbert Spencer",
-    "email": "Jennings_Koss16@yahoo.com",
-    "phone": "859.705.5699 x2444"
-  },
-  {
-    "id": "CKQ20769",
-    "name": "Helen Wisozk",
-    "email": "Tito_Murazik@hotmail.com",
-    "phone": "710.449.7325 x18116"
-  },
-  {
-    "id": "ABA22315",
-    "name": "Sharon Kreiger",
-    "email": "Odell_Bins@yahoo.com",
-    "phone": "531.768.5941 x160"
-  },
-  {
-    "id": "YGG86340",
-    "name": "Loretta Graham",
-    "email": "Larissa.Koelpin78@hotmail.com",
-    "phone": "553.828.1214 x147"
-  },
-  {
-    "id": "WYE27467",
-    "name": "Tomas Dach",
-    "email": "Raquel_OKon@hotmail.com",
-    "phone": "989-349-7665 x208"
-  },
-  {
-    "id": "MIM24306",
-    "name": "Tonya DuBuque",
-    "email": "Jacklyn.Thiel51@gmail.com",
-    "phone": "441-410-6871 x16337"
-  },
-  {
-    "id": "GHD77307",
-    "name": "Jamie Lakin",
-    "email": "Ines_Upton-Miller@yahoo.com",
-    "phone": "1-252-990-4166"
-  }
-]
+[]
