+++ data/testData/index.ts
@@ -1,6 +1,6 @@
-import addresses from './addresses';
-import contacts from './contacts';
-import orgs from './orgs';
+import { schema as addresses } from '@/app/api/addresses/schemas';
+import { schema as contacts } from '@/app/api/contacts/schemas';
+import { schema as orgs } from '@/app/api/orgs/schemas';
 
 const allConfigs = {
   addresses,
