+++ data/testData/orgs.json
@@ -1,114 +1,1 @@
-[
-  {
-    "id": "XFD86998",
-    "name": "Online Card",
-    "website": "online_card.io",
-    "addressId": "BZE58966",
-    "contactIds": [
-      "EBT39485",
-      "RBB15076",
-      "UHD91237"
-    ],
-    "referralIds": "KFZ47645"
-  },
-  {
-    "id": "SAK26742",
-    "name": "Virtual System",
-    "website": "virtual_system.com",
-    "addressId": "QAO87364",
-    "contactIds": [
-      "ZEB62110",
-      "SGW15873"
-    ],
-    "referralIds": "XEZ82429"
-  },
-  {
-    "id": "OMX36774",
-    "name": "Primary Array",
-    "website": "primary_array.js",
-    "addressId": "BZA57636",
-    "contactIds": [
-      "KVI58029",
-      "UQZ77765",
-      "YQC62398"
-    ],
-    "referralIds": "WVD55140"
-  },
-  {
-    "id": "HHS22494",
-    "name": "Multi-byte Microchip",
-    "website": "multi-byte_microchip.com",
-    "addressId": "ZFX88165",
-    "contactIds": [
-      "RUJ47066",
-      "QYA13269"
-    ],
-    "referralIds": "BRH76774"
-  },
-  {
-    "id": "AAS56715",
-    "name": "Neural Matrix",
-    "website": "neural_matrix.io",
-    "addressId": "VKB89750",
-    "contactIds": [
-      "YRB82465",
-      "GFA71209"
-    ],
-    "referralIds": "CKQ20769"
-  },
-  {
-    "id": "LIE18308",
-    "name": "1080p Circuit",
-    "website": "1080p_circuit.com",
-    "addressId": "FRJ95095",
-    "contactIds": [
-      "EEH91210"
-    ],
-    "referralIds": "ABA22315"
-  },
-  {
-    "id": "JNN78913",
-    "name": "Cross-platform Panel",
-    "website": "cross-platform_panel.js",
-    "addressId": "DEN46954",
-    "contactIds": [
-      "LUX21081",
-      "NPE23670",
-      "FLR40888"
-    ],
-    "referralIds": "YGG86340"
-  },
-  {
-    "id": "OML30491",
-    "name": "Wireless Program",
-    "website": "wireless_program.io",
-    "addressId": "SUK12348",
-    "contactIds": [
-      "TNQ85709"
-    ],
-    "referralIds": "WYE27467"
-  },
-  {
-    "id": "EDP87539",
-    "name": "Redundant Sensor",
-    "website": "redundant_sensor.com",
-    "addressId": "BRA29753",
-    "contactIds": [
-      "BIW40830",
-      "KXT29950",
-      "OVZ85021"
-    ],
-    "referralIds": "MIM24306"
-  },
-  {
-    "id": "RUT33550",
-    "name": "Optical Capacitor",
-    "website": "optical_capacitor.js",
-    "addressId": "TUZ29864",
-    "contactIds": [
-      "VKK71026",
-      "KXH13867"
-    ],
-    "referralIds": "GHD77307"
-  }
-]
+[]
